+++ src/components/pages/OrderDetailPage.tsx
@@ -3,14 +3,17 @@ import PageTemplate from '../templates/PageTemplate';
 import { Text } from '../common/Text/Text';
 import { useNavigate, useParams } from 'react-router-dom';
 import useOrderDetail from '../../hooks/useOrderDetail';
-import { skeletonAnimation } from '../box/ProductItem/ProductItem';
 import TextList from '../common/TextList/TextList';
-import Button from '../common/Button/Button';
+import OrderItem from '../box/Order/OrderItem';
+import Modal from '../common/Modal/Modal';
+import ConfirmOrderModal from '../common/Modal/ConfirmOrderModal';
+import DeleteOrderModal from '../common/Modal/DeleteOrderModal';
+import { confirmOrderModalState, deleteOrderModalState } from '../../service/atom';
 
 const OrderDetailPage = () => {
   const { orderId } = useParams();
 
-  const { orderDetailData, confirmOrderAPI, deleteOrderAPI } = useOrderDetail(orderId);
+  const { orderDetailData } = useOrderDetail(orderId);
 
   const navigate = useNavigate();
 
@@ -21,15 +24,9 @@ const OrderDetailPage = () => {
     >
       <CartPageWrapper>
         <CartPageHead style={{ position: 'relative' }}>
-          <div
-            style={{ position: 'absolute', left: 30, cursor: 'pointer' }}
-            onClick={() => navigate('/order')}
-          >
-            <img
-              src="https://cdn.pixabay.com/photo/2020/03/22/15/19/arrow-4957487_1280.png"
-              style={{ width: 28, height: 28 }}
-            />
-          </div>
+          <BackButton onClick={() => navigate('/order')}>
+            <BackButtonImage src="https://cdn.pixabay.com/photo/2020/03/22/15/19/arrow-4957487_1280.png" />
+          </BackButton>
           <Text size="extraLarge" weight="bold">
             주문내역 상세
           </Text>
@@ -37,47 +34,7 @@ const OrderDetailPage = () => {
         <CartPageContent>
           {orderDetailData && (
             <OrderDetailInnerWrapper>
-              <OrderItemWrapper>
-                <OrderItemHead>
-                  <Text size="smaller" weight="light">
-                    주문번호 {orderDetailData.id}
-                  </Text>
-                  {!orderDetailData.confirmState && (
-                    <ButtonWrapper>
-                      <Button
-                        text="주문확정"
-                        size="small"
-                        primary
-                        borderRadius={2}
-                        onClick={() => confirmOrderAPI.mutate()}
-                      />
-                      <Button
-                        text="주문취소"
-                        size="small"
-                        borderRadius={2}
-                        onClick={() => deleteOrderAPI.mutate()}
-                      />
-                    </ButtonWrapper>
-                  )}
-                </OrderItemHead>
-                <OrderProductList>
-                  {orderDetailData.orderProducts.map((orderProduct) => (
-                    <OrderProduct key={orderProduct.product.id}>
-                      <ProductImage src={orderProduct.product.imageUrl} />
-                      <ProductInfo>
-                        <Text size="small" weight="normal">
-                          {orderProduct.product.name}
-                        </Text>
-                        <Text size="smaller" weight="light" color="rgba(0,0,0,0.4)">
-                          {`${orderProduct.product.price.toLocaleString()}원 / ${
-                            orderProduct.quantity
-                          }개`}
-                        </Text>
-                      </ProductInfo>
-                    </OrderProduct>
-                  ))}
-                </OrderProductList>
-              </OrderItemWrapper>
+              <OrderItem order={orderDetailData} />
               <OrderDetailPriceBox>
                 <BoxHead>
                   <Text size="smaller" weight="light">
@@ -121,11 +78,28 @@ const OrderDetailPage = () => {
           )}
         </CartPageContent>
       </CartPageWrapper>
+      <Modal modalState={confirmOrderModalState}>
+        <ConfirmOrderModal />
+      </Modal>
+      <Modal modalState={deleteOrderModalState}>
+        <DeleteOrderModal />
+      </Modal>
     </PageTemplate>
   );
 };
 
 export default OrderDetailPage;
+
+const BackButton = styled.div`
+  position: absolute;
+  left: 30px;
+  cursor: pointer;
+`;
+
+const BackButtonImage = styled.img`
+  width: 28px;
+  height: 28px;
+`;
 
 const CartPageWrapper = styled.div`
   width: 100%;
@@ -174,62 +148,6 @@ const CartPageContent = styled.div`
   }
 `;
 
-const OrderItemWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  border-radius: 6px;
-`;
-
-const OrderItemHead = styled.div`
-  border-radius: 6px 6px 0 0;
-  width: 100%;
-  height: 62px;
-  background-color: rgb(243, 245, 247);
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-  padding: 0 26px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const OrderProductList = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-`;
-
-const OrderProduct = styled.div`
-  padding: 40px 26px;
-  display: flex;
-
-  &:not(:last-child) {
-    border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-  }
-`;
-
-const ProductImage = styled.img`
-  width: 141px;
-  height: 141px;
-  object-fit: cover;
-  margin-right: 44px;
-  border-radius: 3px;
-  transition: all 0.32s ease;
-  border: 1px solid rgba(0, 0, 0, 0.05);
-
-  background: linear-gradient(120deg, #e5e5e5 20%, #f0f0f0 28%, #f0f0f0 40%, #e5e5e5 48%);
-  background-position: 100% 0;
-  background-size: 282px;
-  animation: ${skeletonAnimation} 1s infinite;
-`;
-
-const ProductInfo = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-`;
-
 const OrderDetailInnerWrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -267,10 +185,4 @@ const TotalPriceWrapper = styled.div`
   margin: 12px 0 0 0;
   padding: 12px 0 0 0;
   border-top: 1px solid rgba(0, 0, 0, 0.1);
-`;
-
-const ButtonWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 10px;
 `;
